% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_UniFrac_fork.R
\name{fast_unweighted_UniFrac}
\alias{fast_unweighted_UniFrac}
\title{Fast Unweighted UniFrac}
\usage{
fast_unweighted_UniFrac(tips_abun, tree, ncores = 1)
}
\arguments{
\item{tips_abun}{data.frame object containing taxonomic abundances across samples, with rows as taxa and columns as samples.}

\item{tree}{phylo object, which must contain all row names of \code{tips_abun} as tip labels.}

\item{ncores}{integer indicating number of cores to use for parallelizable steps.}
}
\description{
Compute pairwise unweighted UniFrac distances between samples based on taxa abundances and input tree. Returns matrix with distances in upper triangle.
}
\details{
This is a modified version of the fastUniFrac function taken from v1.39.1 of phyloseq (https://github.com/joey711/phyloseq).
This code was distributed under the AGPL-3 license: https://www.gnu.org/licenses/agpl-3.0.en.html.
function taken from v1.39.1 of phyloseq (https://github.com/joey711/phyloseq).
This code was distributed under the AGPL-3 license: https://www.gnu.org/licenses/agpl-3.0.en.html.

Keys changes:
\itemize{
\item Separate functions for weighted vs unweighted UniFrac
\item Distances are always normalized to be within 0 - 1 (i.e., the "normalized" option was removed).
\item They are compatible with non-phyloseq input objects.
\item The parallel package (and mclapply) is used rather than foreach, just to be compatible with the rest of FuncDiv.
\item Matrix of distances rather than a dist object will be returned.
}

The original function was adapted by the phyloseq authors based on
The ISME Journal (2010) 4, 17-27; doi:10.1038/ismej.2009.97;
http://www.nature.com/ismej/journal/v4/n1/full/ismej200997a.html
}
