% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{func_abun_crossproduct}
\alias{func_abun_crossproduct}
\title{Utility function to get community-wide function abundance table}
\usage{
func_abun_crossproduct(func_tab, abun_tab)
}
\arguments{
\item{func_tab}{data.frame object containing function copy numbers, with rows as functions and columns as taxa.}

\item{abun_tab}{data.frame object containing taxonomic abundances across samples, with rows as taxa and columns as samples.}
}
\value{
data.frame representing the \emph{unnormalized} community-wide abundances of functions across samples.
}
\description{
Takes in table of function copy numbers across taxa and table of taxa abundances across samples.
I.e., it represents the multiplication of the function copy numbers by the abundances of the taxa within each sample.
}
