% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{create_jensen_shannon_divergence_FuncPtr}
\alias{create_jensen_shannon_divergence_FuncPtr}
\title{Function to create Rcpp pointer for Jensen-Shannon divergence function}
\format{
\subsection{\code{create_jensen_shannon_divergence_FuncPtr()}}{

Function that a returns a RcppXPtrUtils::cppXPtr pointer.
Call without parentheses to see the code: \code{create_jensen_shannon_divergence_FuncPtr}
}
}
\usage{
create_jensen_shannon_divergence_FuncPtr
}
\description{
ParallelDist can take custom distance metrics in as Rcpp pointers. However,
this was causing issues when defined as a normal R object in the FuncDiv
package, so instead, the pointer will be created as needed internally with
this packaged function.

This also serves as an example of how one can specify other custom pointer
functions to compute beta diversity (see below).
}
\keyword{datasets}
